♻️ refactor(entities): update PurchaseOrderEntity ID fields to int64
- Changed `ID`, `BuyerID`, `CarrierID`, `OrderStatusID`, and `WarehouseID` from `int` to `int64`.
- Improves database compatibility for handling large integer IDs.

+++ internal/infrastructure/repository/entities/purchase_order_entity.go
@@ -1,14 +1,14 @@
 package entities
 
 type PurchaseOrderEntity struct {
-	ID            int
+	ID            int64
 	OrderNumber   string
 	OrderDate     string
 	TrackingCode  string
-	BuyerID       int
-	CarrierID     int
-	OrderStatusID int
-	WarehouseID   int
+	BuyerID       int64
+	CarrierID     int64
+	OrderStatusID int64
+	WarehouseID   int64
 }
 
 func (p *PurchaseOrderEntity) GetByIdQuery(id int) (query string, args []interface{}) {
